+++ packages/contracts/test/aggregator/AggregatorWithIPCreationProxy.spec.ts
@@ -1,4 +1,4 @@
-import './setup'
+import '../setup'
 
 /* External Imports */
 import { createMockProvider, deployContract, getWallets } from 'ethereum-waffle'
@@ -6,8 +6,8 @@ import chai = require('chai')
 const { expect } = chai
 
 /* Contract imports */
-import * as AggregatorWithIPCreationProxy from '../build/AggregatorWithIPCreationProxy.json'
-import * as AggregatorRegistry from '../build/AggregatorRegistry.json'
+import * as AggregatorWithIPCreationProxy from '../../build/AggregatorWithIPCreationProxy.json'
+import * as AggregatorRegistry from '../../build/AggregatorRegistry.json'
 
 /* Begin tests */
 describe('AggregatorWithIPCreationProxy', () => {
